+++ src/onegov.election_day/onegov/election_day/models/archive.py
@@ -28,13 +28,46 @@ class Archive(object):
             lambda i: (i.__class__.__name__.lower(), i.domain, i.date)
         )
 
-    def latest(self):
+    def to_json(self, items, request):
+        results = []
+        for item in items:
+            values = {
+                'data_url': request.link(item, 'json'),
+                'date': item.date.isoformat(),
+                'domain': item.domain,
+                'progress': {},
+                'title': item.title_translations,
+                'type': item.__class__.__name__.lower(),
+                'url': request.link(item),
+            }
+
+            try:
+                values['progress']['counted'] = (
+                    item.counted_municipalities or 0
+                )
+                values['progress']['total'] = item.total_municipalities or 0
+            except AttributeError:
+                divider = len(item.ballots) or 1
+                values['progress']['counted'] = item.progress[0] or 0 / divider
+                values['progress']['total'] = item.progress[1] or 0 / divider
+                values['answer'] = item.answer or ""
+                values['yeas_percentage'] = item.yeas_percentage
+                values['nays_percentage'] = item.nays_percentage
+
+            results.append(values)
+
+        return results
+
+    def latest(self, group=True):
         items = ElectionCollection(self.session).get_latest() or []
         items.extend(VoteCollection(self.session).get_latest() or [])
 
-        return self.group_items(items, reverse=True)
+        return self.group_items(items, reverse=True) if group else items
 
-    def by_date(self):
+    def latest_json(self, request):
+        return self.to_json(self.latest(group=False), request)
+
+    def by_date(self, group=True):
         try:
             _date = date.fromtimestamp(mktime(strptime(self.date, '%Y-%m-%d')))
             items = ElectionCollection(self.session).by_date(_date) or []
@@ -43,4 +76,7 @@ class Archive(object):
             items = ElectionCollection(self.session).by_year(self.date) or []
             items.extend(VoteCollection(self.session).by_year(self.date) or [])
 
-        return self.group_items(items)
+        return self.group_items(items) if group else items
+
+    def by_date_json(self, request):
+        return self.to_json(self.by_date(group=False), request)