+++ src/onegov.election_day/onegov/election_day/tests/test_views.py
@@ -209,6 +209,58 @@ def test_opendata(election_day_app):
     assert "Open Data" in client.get('/opendata')
 
 
+def test_view_latest(election_day_app):
+    client = Client(election_day_app)
+    client.get('/locale/de_CH').follow()
+
+    login(client)
+
+    new = client.get('/manage/new-vote')
+    new.form['vote_de'] = "Abstimmung 1. Januar 2013"
+    new.form['date'] = date(2013, 1, 1)
+    new.form['domain'] = 'federation'
+    new.form.submit()
+
+    new = client.get('/manage/new-election')
+    new.form['election_de'] = "Wahl 1. Januar 2013"
+    new.form['date'] = date(2013, 1, 1)
+    new.form['mandates'] = 1
+    new.form['election_type'] = 'majorz'
+    new.form['domain'] = 'federation'
+    new.form.submit()
+
+    latest = client.get('/')
+    assert "Abstimmung 1. Januar 2013" in latest
+    assert "Wahl 1. Januar 2013" in latest
+
+
+def test_view_latest_json(election_day_app):
+    client = Client(election_day_app)
+    client.get('/locale/de_CH').follow()
+
+    assert client.get('/json').json['results'] == []
+
+    login(client)
+
+    new = client.get('/manage/new-vote')
+    new.form['vote_de'] = "Abstimmung 1. Januar 2013"
+    new.form['date'] = date(2013, 1, 1)
+    new.form['domain'] = 'federation'
+    new.form.submit()
+
+    new = client.get('/manage/new-election')
+    new.form['election_de'] = "Wahl 1. Januar 2013"
+    new.form['date'] = date(2013, 1, 1)
+    new.form['mandates'] = 1
+    new.form['election_type'] = 'majorz'
+    new.form['domain'] = 'federation'
+    new.form.submit()
+
+    latest = client.get('/json')
+    assert "Abstimmung 1. Januar 2013" in latest
+    assert "Wahl 1. Januar 2013" in latest
+
+
 def test_view_archive(election_day_app):
     client = Client(election_day_app)
     client.get('/locale/de_CH').follow()
@@ -241,6 +293,38 @@ def test_view_archive(election_day_app):
 
     archive = client.get('/archive/2013-02-02')
     assert "noch keine Wahlen oder Abstimmungen" in archive
+
+
+def test_view_archive_json(election_day_app):
+    client = Client(election_day_app)
+    client.get('/locale/de_CH').follow()
+
+    login(client)
+
+    new = client.get('/manage/new-vote')
+    new.form['vote_de'] = "Abstimmung 1. Januar 2013"
+    new.form['date'] = date(2013, 1, 1)
+    new.form['domain'] = 'federation'
+    new.form.submit()
+
+    new = client.get('/manage/new-election')
+    new.form['election_de'] = "Wahl 1. Januar 2013"
+    new.form['date'] = date(2013, 1, 1)
+    new.form['mandates'] = 1
+    new.form['election_type'] = 'majorz'
+    new.form['domain'] = 'federation'
+    new.form.submit()
+
+    archive = client.get('/archive/2013/json')
+    assert "Abstimmung 1. Januar 2013" in archive
+    assert "Wahl 1. Januar 2013" in archive
+
+    archive = client.get('/archive/2013-01-01/json')
+    assert "Abstimmung 1. Januar 2013" in archive
+    assert "Wahl 1. Januar 2013" in archive
+
+    archive = client.get('/archive/2013-02-02/json')
+    assert archive.json['results'] == []
 
 
 def test_view_election_candidates(election_day_app_gr):